Fixing calculations using doubles, have to cast double to be able to calculate exactly. TODO A way to keep track of different shifts. Figure out when updating to clear all other information that is not needed, like cash received. When entering dollar amount, automatically position the two decimal numbers.

+++ app/src/main/java/com/digitalruiz/pizzadriver/AddOrder.java
@@ -428,7 +428,7 @@ public class AddOrder extends AppCompatActivity {
                 else {
                     CashReceived = Double.parseDouble(cashReceivedEditText.getText().toString());
                 }
-                double tip = CashReceived - OrderTotal;
+                double tip = (double) CashReceived - OrderTotal;
                 Log.v("Test", "changed" + tip);
                 tipEditText.setText(Double.toString(tip));
             }
@@ -465,7 +465,7 @@ public class AddOrder extends AppCompatActivity {
                 else {
                     CashReceived = Double.parseDouble(cashReceivedEditText.getText().toString());
                 }
-                double tip = CashReceived - OrderTotal;
+                double tip = (double) CashReceived - OrderTotal;
                 Log.v("Test", "changed" + tip);
                 tipEditText.setText(Double.toString(tip));
             }
